refactror: 부동산페이지 api호출 query파일에 병합
